Create Number_Guessing_Game.py
'''The objective of this project is to build a simple number guessing game that challenges the user to identify a randomly
selected number within a specified range. The game begins by allowing the user to define a range by entering a lower and an 
upper bound (for example, from A to B). Once the range is set, the system randomly selects an integer that falls within this 
user-defined interval. The user's task is then to guess the chosen number using as few attempts as possible. The game provides 
feedback after each guess, helping the user refine their next guess based on whether their previous attempt was too high or
too low.''